Time: 0 ms (100%), Space: 43.3 MB (67.02%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.java
@@ -15,9 +15,10 @@
  */
 class Solution {
     public boolean isValidBST(TreeNode root) {
-        return validate(root, Integer.MIN_VALUE, Integer.MAX_VALUE);
+
+        return validate(root, Long.MIN_VALUE, Long.MAX_VALUE);
     }
-    boolean validate(TreeNode root, int min, int max){
+    boolean validate(TreeNode root, long min, long max){
         if(root == null) return true;
         if(root.val >= max || root.val <= min) return false;
         return validate(root.left, min, root.val) && validate(root.right, root.val, max);
